🐞 fix: 浏览器关闭函数逻辑整改修复

+++ lib/functions/assignment_notifier/bg_worker.dart
@@ -38,7 +38,7 @@ class AssignmentNotifierBgWorker {
   static void forceStop() async {
     try {
       AssignmentNotifierBgWorker.bgWorker.cancel();
-      await AssignmentNotifierBgWorker.closeBrowser();
+      AssignmentNotifierBgWorker.closeBrowser();
     } catch (_) {}
   }
 
@@ -48,13 +48,14 @@ class AssignmentNotifierBgWorker {
     isar.writeTxn(() => isar.jupiterDatas.put(jupiterData));
   }
 
-  static Future<void> closeBrowser() async {
-    try {
-      browser.close();
-      await Future.delayed(const Duration(milliseconds: 320));
-      Utils.shell.run("taskkill /f /t /im chrome.exe");
-      await Future.delayed(const Duration(seconds: 2));
-    } catch (_) {}
+  static void closeBrowser() async {
+    browser.close();
+    await Future.delayed(const Duration(seconds: 5));
+    if (!browser.isConnected) {
+      try {
+        Utils.shell.run("taskkill /f /t /im chrome.exe");
+      } catch (_) {}
+    }
   }
 
   static Future<Page?> openJupiterPage() async {
@@ -108,7 +109,7 @@ class AssignmentNotifierBgWorker {
 
       Log.logger.i("浏览器启动");
     } catch (e) {
-      await closeBrowser();
+      closeBrowser();
       Log.logger.e("browserClose".tr, error: e);
       final error = "Chromium 自动化浏览器启动异常: $e";
       dataFetchStatus.value = "-$error";
@@ -126,7 +127,7 @@ class AssignmentNotifierBgWorker {
         break;
       } catch (e) {
         if (timesOfErr > 2) {
-          await closeBrowser();
+          closeBrowser();
           Log.logger.e("browserClose".tr, error: e);
           final error = "网络错误 或 请求被拦截(请修改设置中的 Cloudflare Bypass Cookies): $e";
           dataFetchStatus.value = "-$error";
@@ -151,7 +152,7 @@ class AssignmentNotifierBgWorker {
       await Future.delayed(Constants.universalDelay);
       await jupiterPage.click("#menulist_region1 > div[val='xx_xx']");
     } catch (e) {
-      await closeBrowser();
+      closeBrowser();
       Log.logger.e("browserClose".tr, error: e);
       final error = "登录 Jupiter 时发生未知异常: $e";
       dataFetchStatus.value = "-$error";
@@ -166,7 +167,7 @@ class AssignmentNotifierBgWorker {
         await jupiterPage.type("#text_password1", password);
         await jupiterPage.click("#loginbtn");
       } catch (e) {
-        await closeBrowser();
+        closeBrowser();
         Log.logger.e("browserClose".tr, error: e);
         final error = "登录 Jupiter 时发生未知异常: $e";
         dataFetchStatus.value = "-$error";
@@ -179,7 +180,7 @@ class AssignmentNotifierBgWorker {
         await jupiterPage.type("#text_password1", SettingManager.settings["jupiterPassword"]!);
         await jupiterPage.click("#loginbtn");
       } catch (e) {
-        await closeBrowser();
+        closeBrowser();
         Log.logger.e("browserClose".tr, error: e);
         final error = "登录 Jupiter 时发生未知异常: $e";
         dataFetchStatus.value = "-$error";
@@ -194,12 +195,12 @@ class AssignmentNotifierBgWorker {
 
       // 用于特殊调用，直接检查账号是否可用
       if (name != null && password != null) {
-        await closeBrowser();
+        closeBrowser();
         Log.logger.i("browserClose".tr);
         return true;
       }
     } catch (_) {
-      await closeBrowser();
+      closeBrowser();
       const error = "Jupiter 账号或密码错误";
       dataFetchStatus.value = "-$error";
       UI.showNotification(error, type: NotificationType.error);
@@ -213,7 +214,7 @@ class AssignmentNotifierBgWorker {
       await Future.delayed(Constants.universalDelay);
       await jupiterPage.click("#schoolyearlist > div:nth-child(1)");
     } catch (e) {
-      await closeBrowser();
+      closeBrowser();
       Log.logger.e("browserClose".tr, error: e);
       const error = "学年列表出现异常，无法正常选择最新学年";
       dataFetchStatus.value = "-$error";
@@ -227,7 +228,7 @@ class AssignmentNotifierBgWorker {
   static void checkForNewAssignment({int? retry}) async {
     // 如果传入的 retry 为 0 直接终止
     if (retry == 0) {
-      await closeBrowser();
+      closeBrowser();
       Log.logger.i("browserClose".tr);
       return;
     }
@@ -245,12 +246,12 @@ class AssignmentNotifierBgWorker {
       courseCount = (await getCourses(jupiterPage)).length;
       await jupiterPage.click("#touchnavbtn");
     } catch (_) {
-      await closeBrowser();
+      closeBrowser();
       return;
     }
 
     if (courseCount == 0) {
-      await closeBrowser();
+      closeBrowser();
       lastUpdateTime.value = Utils.formatTime(DateTime.now());
       dataFetchStatus.value = "+";
       Log.logger.i("browserClose".tr);
@@ -380,14 +381,14 @@ class AssignmentNotifierBgWorker {
       final error = "Chromium 自动化浏览器出现上下文异常，将进行重试: $e";
       dataFetchStatus.value = "-$error";
       UI.showNotification(error, type: NotificationType.error);
-      await closeBrowser();
+      closeBrowser();
       checkForNewAssignment(retry: retry == null ? 3 : --retry);
       return;
     }
 
     // 如无修改就不再写入数据库，反之写入
     if (modification == 0) {
-      await closeBrowser();
+      closeBrowser();
       lastUpdateTime.value = Utils.formatTime(DateTime.now());
       dataFetchStatus.value = "+";
       Log.logger.i("browserClose".tr);
@@ -422,7 +423,7 @@ class AssignmentNotifierBgWorker {
           final error = "Chromium 自动化浏览器出现上下文异常，作业详情信息获取失败: $e";
           dataFetchStatus.value = "-$error";
           UI.showNotification(error, type: NotificationType.error);
-          await closeBrowser();
+          closeBrowser();
         }
 
         final course = jupiterData.courses!.firstWhere((course) => (course.name == courseName));
@@ -433,7 +434,7 @@ class AssignmentNotifierBgWorker {
       isar.writeTxn(() => isar.jupiterDatas.put(jupiterData));
     }
 
-    await closeBrowser();
+    closeBrowser();
     dataFetchStatus.value = "+";
     Log.logger.i("browserClose".tr);
   }
@@ -464,7 +465,7 @@ class AssignmentNotifierBgWorker {
         await Future.delayed(Constants.universalDelay);
       } catch (e) {
         if (timesOfErr > 3) {
-          await closeBrowser();
+          closeBrowser();
           Log.logger.e("browserClose".tr, error: e);
           final error = "Chromium 自动化浏览器出现多次上下文异常，作业详情信息获取失败: $e";
           dataFetchStatus.value = "-$error";
